@@ -18,7 +18,7 @@
       "test": "vendor/bin/phpunit"
     },
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^8.0",
         "ext-pdo": "*"
     },
     "require-dev": {
