@@ -28,19 +28,19 @@ class HavingTest extends BaseClass
     public function testColumn()
     {
         $expected = 'SELECT * FROM "users" GROUP BY "age" HAVING COUNT("friends") > 5';
-        $actual = $this->db->from('users')
+        $this->db->from('users')
             ->groupBy('age')
             ->having('friends', function (HavingExpression $column) {
                 $column->count()->gt(5);
             })
             ->select();
-        $this->assertEquals($expected, $actual);
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testExpression()
     {
         $expected = 'SELECT * FROM "users" GROUP BY "age" HAVING COUNT("friends" * 2) > 5';
-        $actual = $this->db->from('users')
+        $this->db->from('users')
             ->groupBy('age')
             ->having(function (Expression $expr) {
                 $expr->column('friends')->{'*'}->value(2);
@@ -48,13 +48,13 @@ class HavingTest extends BaseClass
                 $column->count()->gt(5);
             })
             ->select();
-        $this->assertEquals($expected, $actual);
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testNested()
     {
-        $actual = 'SELECT COUNT("orders"."id") AS "total_orders", "customers"."name" AS "name" FROM "customers" LEFT JOIN "orders" ON "customers"."id" = "orders"."cid" GROUP BY LCASE("customers"."name") HAVING COUNT("orders"."id") > 10 AND (SUM("orders"."value") >= 1000 OR MIN(ROUND("orders"."value", 2)) >= 500)';
-        $expected = $this->db->from('customers')
+        $expected = 'SELECT COUNT("orders"."id") AS "total_orders", "customers"."name" AS "name" FROM "customers" LEFT JOIN "orders" ON "customers"."id" = "orders"."cid" GROUP BY LCASE("customers"."name") HAVING COUNT("orders"."id") > 10 AND (SUM("orders"."value") >= 1000 OR MIN(ROUND("orders"."value", 2)) >= 500)';
+        $this->db->from('customers')
             ->leftJoin('orders', function(Join $join){
                 $join->on('customers.id', 'orders.cid');
             })
@@ -78,6 +78,6 @@ class HavingTest extends BaseClass
                 $include->count('orders.id', 'total_orders')
                     ->column('customers.name', 'name');
             });
-        $this->assertEquals($expected, $actual);
+        $this->assertEquals($expected, $this->getSQL());
     }
 }
