@@ -24,32 +24,32 @@ class InsertTest extends BaseClass
     public function testInsertSingleValue()
     {
         $expected = 'INSERT INTO "users" ("age") VALUES (18)';
-        $actual = $this->db->insert(['age' => 18])->into('users');
-        $this->assertEquals($expected, $actual);
+        $this->db->insert(['age' => 18])->into('users');
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testInsertMultipleValues()
     {
         $expected = 'INSERT INTO "users" ("name", "age") VALUES (\'foo\', 18)';
-        $actual = $this->db->insert(['name' => 'foo', 'age' => 18])->into('users');
-        $this->assertEquals($expected, $actual);
+        $this->db->insert(['name' => 'foo', 'age' => 18])->into('users');
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testInsertBooleanValues()
     {
         $expected = 'INSERT INTO "test" ("foo", "bar") VALUES (TRUE, FALSE)';
-        $actual = $this->db->insert(['foo' => true, 'bar' => false])->into('test');
-        $this->assertEquals($expected, $actual);
+        $this->db->insert(['foo' => true, 'bar' => false])->into('test');
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testInsertExpressions()
     {
         $expected = 'INSERT INTO "users" ("name") VALUES (LCASE( \'foo\' ))';
-        $actual = $this->db->insert([
+        $this->db->insert([
             'name' => function (Expression $expr) {
                 $expr->{'LCASE('}->value('foo')->{')'};
             },
         ])->into('users');
-        $this->assertEquals($expected, $actual);
+        $this->assertEquals($expected, $this->getSQL());
     }
 }
