@@ -22,15 +22,15 @@ class LimitTest extends BaseClass
     public function testLimit()
     {
         $expected = 'SELECT * FROM "users" ORDER BY "name" ASC LIMIT 25';
-        $actual = $this->db->from('users')->orderBy('name')->limit(25)->select();
-        $this->assertEquals($expected, $actual);
+        $this->db->from('users')->orderBy('name')->limit(25)->select();
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testOffset()
     {
         $expected = 'SELECT * FROM "users" ORDER BY "name" ASC LIMIT 25 OFFSET 10';
-        $actual = $this->db->from('users')->orderBy('name')->limit(25)->offset(10)->select();
-        $this->assertEquals($expected, $actual);
+        $this->db->from('users')->orderBy('name')->limit(25)->offset(10)->select();
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     /*
