@@ -24,59 +24,59 @@ class SelectAggregateTest extends BaseClass
     public function testCountNoColumns()
     {
         $expected = 'SELECT COUNT(*) FROM "users"';
-        $actual = $this->db->from('users')->count();
-        $this->assertEquals($expected, $actual);
+        $this->db->from('users')->count();
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testCountOneColumn()
     {
         $expected = 'SELECT COUNT("description") FROM "users"';
-        $actual = $this->db->from('users')->count('description');
-        $this->assertEquals($expected, $actual);
+        $this->db->from('users')->count('description');
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testCountOneColumnDistinct()
     {
         $expected = 'SELECT COUNT(DISTINCT "description") FROM "users"';
-        $actual = $this->db->from('users')->count('description', true);
-        $this->assertEquals($expected, $actual);
+        $this->db->from('users')->count('description', true);
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testLargestValue()
     {
         $expected = 'SELECT MAX("age") FROM "users"';
-        $actual = $this->db->from('users')->max('age');
-        $this->assertEquals($expected, $actual);
+        $this->db->from('users')->max('age');
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testSmallestValue()
     {
         $expected = 'SELECT MIN("age") FROM "users"';
-        $actual = $this->db->from('users')->min('age');
-        $this->assertEquals($expected, $actual);
+        $this->db->from('users')->min('age');
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testAverageValue()
     {
         $expected = 'SELECT AVG("age") FROM "users"';
-        $actual = $this->db->from('users')->avg('age');
-        $this->assertEquals($expected, $actual);
+        $this->db->from('users')->avg('age');
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testTotalSum()
     {
         $expected = 'SELECT SUM("age") FROM "users"';
-        $actual = $this->db->from('users')->sum('age');
-        $this->assertEquals($expected, $actual);
+        $this->db->from('users')->sum('age');
+        $this->assertEquals($expected, $this->getSQL());
     }
 
     public function testExpressionAggregate()
     {
         $expected = 'SELECT SUM("friends" - "enemies") FROM "users"';
-        $actual = $this->db->from('users')->sum(function (Expression $expr) {
+        $this->db->from('users')->sum(function (Expression $expr) {
             $expr->column('friends')->{'-'}->column("enemies");
         });
-        $this->assertEquals($expected, $actual);
+        $this->assertEquals($expected, $this->getSQL());
     }
 
 }
