@@ -57,35 +57,35 @@ class BaseClass extends TestCase
 
     public function testCreateTable()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
 
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testAddSingleColumn()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testAddMultipleColumns()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a');
             $table->integer('b');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testTypes()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a');
             $table->float('b');
             $table->double('c');
@@ -105,12 +105,12 @@ class BaseClass extends TestCase
             $table->text('m');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testIntSizes()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a')->size('tiny');
             $table->integer('b')->size('small');
             $table->integer('c')->size('normal');
@@ -118,12 +118,12 @@ class BaseClass extends TestCase
             $table->integer('e')->size('big');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testTextSizes()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->text('a')->size('tiny');
             $table->text('b')->size('small');
             $table->text('c')->size('normal');
@@ -131,12 +131,12 @@ class BaseClass extends TestCase
             $table->text('e')->size('big');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testBinarySizes()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->binary('a')->size('tiny');
             $table->binary('b')->size('small');
             $table->binary('c')->size('normal');
@@ -144,61 +144,61 @@ class BaseClass extends TestCase
             $table->binary('e')->size('big');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testColumnProperties()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a')->unsigned();
             $table->float('b')->defaultValue(0.1);
             $table->string('c')->notNull();
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testColumnConstraints()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a')->primary();
             $table->integer('b')->unique();
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testColumnNamedConstraints()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a')->primary('pk_a');
             $table->integer('b')->unique('uk_b');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testAutoincrement()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a')->autoincrement();
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testNamedAutoincrement()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a')->autoincrement('x');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testIndex()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a')->index();
             $table->integer('b')->index('x');
             $table->integer('c');
@@ -210,12 +210,12 @@ class BaseClass extends TestCase
             $table->index(['c', 'd'], 'z');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testForeignKey()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a');
             $table->foreign('a')
                 ->references('bar', 'a')
@@ -223,12 +223,12 @@ class BaseClass extends TestCase
                 ->onDelete('cascade');
         });;
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testForeignKeyMultiple()
     {
-        $result = $this->schema->create('foo', function (CreateTable $table) {
+        $this->schema->create('foo', function (CreateTable $table) {
             $table->integer('a');
             $table->integer('b');
             $table->foreign(['a', 'b'])
@@ -237,44 +237,44 @@ class BaseClass extends TestCase
                 ->onDelete('cascade');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testAlterTableAddColumn()
     {
-        $result = $this->schema->alter('foo', function(AlterTable $table){
+        $this->schema->alter('foo', function(AlterTable $table){
             $table->integer('a');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testAlterTableAddMultipleColumns()
     {
-        $result = $this->schema->alter('foo', function(AlterTable $table){
+        $this->schema->alter('foo', function(AlterTable $table){
             $table->integer('a');
             $table->integer('b');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testAlterTableDropColumn()
     {
-        $result = $this->schema->alter('foo', function(AlterTable $table){
+        $this->schema->alter('foo', function(AlterTable $table){
             $table->dropColumn('a');
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     public function testAlterTableAddDefaults()
     {
-        $result = $this->schema->alter('foo', function(AlterTable $table){
+        $this->schema->alter('foo', function(AlterTable $table){
             $table->setDefaultValue('a', 100);
         });
 
-        $this->execTest(__FUNCTION__, $result);
+        $this->execTest(__FUNCTION__, $this->schema->getResult());
     }
 
     private function execTest($test, $result)
